+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.27.0
+### Changes
+* Updated readme dependency documentation
+* Added [menu option](src#menu-options) for enabling and disabling the menu confirmation prompt.
+* Changed `? Save selection › (y/N)` prompt, to `? Are you sure › (Y/n)` prompt. Use the [Menu save default script](src#menu-save-default-script) for saving your menu selection.
+
 ## 1.26.3
 ### Fixes
 * Debug logging fix related to change 1.26.2